remove a network request in favor of direct querying
