+++ src/util/post.ts
@@ -203,3 +203,47 @@ export const getRandomPostForSlug = async (
 
   return {};
 };
+
+export const getPostForSlug = async (
+  slug: string | null,
+  origin?: string,
+): Promise<any> => {
+  const allPosts = Object.keys(Posts);
+  const post = allPosts
+    .map((key) => {
+      const component = (Posts as any)[key] as Post;
+      return {
+        title: component.title,
+        slug: component.slug,
+        tags: component.tags ?? [],
+        description: component.description,
+        cover: component.cover,
+        published: component.published,
+        wordCount: (wordCounts as Record<string, number>)[component.slug],
+      };
+    })
+    .find((post) => post.slug === slug);
+
+  if (post) {
+    const conn = connect(config);
+
+    const results = await conn.execute(
+      'SELECT views FROM post_page_views_aggregate WHERE slug = ?',
+      [post.slug],
+    );
+
+    console.log('fetched single post', slug);
+
+    const metadata = getImageMetadata(post.cover);
+
+    return {
+      ...post,
+      viewCount:
+        (results.rows[0] as Record<'views', number> | undefined)?.views ?? 0,
+      metadata,
+      url: `${origin}/blog/${post.slug}`,
+    };
+  }
+
+  return {};
+};
